Move of QStruct.h and inputFileInstance.h into retractor folder
