+++ src/retractor/QStruct.h
